[FEAT] 화살 쏘기 구현 하지만 위치 조정 필요 #64

+++ Client/Header/Player.h
@@ -145,7 +145,7 @@ private:
 
 	_vec3			_pulsepos;
 	float			_attackDelay;
-	float			_arrowCount;
+	int				_arrowCount;
 
 	GameObject*			_inventory[8];
 	Bow*				_weaponSlot[4];
